centralize isMobile logic in store & unify route bar usage
- Moved isMobile detection to App.vue (resize event) and set store.isMobile
- Removed duplicate resize watchers from HomepageView.vue
- Called routeService methods (fitToCurrentRoute/removeRoute) directly
- Ensured single app.use(pinia) + initRouteService in main.js
